requirements.txt
tensorflow==1.14.0
bert_for_tf2==0.3.1
numpy==1.16.4
six==1.12.0
